requirements.txt
fastapi
uvicorn[standard]
orjson
yfinance
pandas
scikit-learn
newsapi-python
nltk
setuptools
numpy==1.26.4
huggingface_hub
tensorflow